Add all ABIs to all data sources
This is a good idea because the mappings of all these data sources
import code like `domain.ts` that in turn imports generated code like
`GenericScheme.ts`. If any of these mappings (by accident or not) start
using e.g. `GenericScheme` in the future without including the
`GenericScheme` ABI, the subgraph would fail.